separated functions for preprocessing, embedding, indexing and topic modeling into different files and implemented them in case_processing
